Create Module 5
For my code review, I looked at the parseAndAddAnimal() method in my ZooManager class. At first, it was kind of messy and relied a lot on splitting strings and guessing where certain words would be, which made it easy to break if the input wasn’t perfect. I went back and added some checks to make sure the data was actually there before trying to use it, trimmed extra spaces, and made a helper method to clean up the long switch statement. I also added some backup options in case the input was weird or missing stuff. After fixing it up, the code runs smoother and is easier to read, and I feel better about handling unexpected input.